refactor(orders): split work_order_target_prechecks into single-concern libs (#4109)
Wave 5 slice B decomposes the 414-line precheck monolith into purchase,
improvement, explore, and shared modules with a thin registry barrel.
