+++ packages/colonizethis_orders/lib/src/orders/validators/work_order_target_prechecks.dart
@@ -1,365 +1,16 @@
-import 'package:colonizethis_data/colonizethis_data.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 
-import '../build_rail_work_rules.dart';
 import '../order_work_constants.dart';
-import '../diplomatic_access_helpers.dart';
-import '../unit_type_helpers.dart';
-import 'package:colonizethis_world/colonizethis_world.dart';
-import 'package:colonizethis_diplomacy/colonizethis_diplomacy.dart';
 import '../order_validation_result.dart';
+import 'work_order_target_prechecks_explore.dart';
+import 'work_order_target_prechecks_improvement.dart';
+import 'work_order_target_prechecks_purchase.dart';
+import 'work_order_target_prechecks_shared.dart';
 
-/// Shared inputs for per-target work order validation steps that run before
-/// generic territory and cost checks. [civilianEmbassyWorkAllowed] mirrors
-/// [WorkOrderValidator] embassy/minor rules for Builder/Engineer/Merchant.
-class WorkOrderTargetPrecheckContext {
-  WorkOrderTargetPrecheckContext({
-    required this.game,
-    required this.player,
-    required this.playerId,
-    required this.treasury,
-    required this.civilianEmbassyWorkAllowed,
-    required this.devExclusiveTiles,
-    this.factionMembership,
-    this.tileMapByRegion,
-  });
-
-  final Game game;
-  final Player player;
-  final String playerId;
-  final int treasury;
-
-  /// Tiles already reserved by accepted dev-exclusive work for this player.
-  final Set<String> devExclusiveTiles;
-
-  /// Per-region tile maps, used to resolve terrain for terrain-dependent caps
-  /// (e.g. the scrub-forest timber level-1 hard cap, R4 / issue #3573). May be
-  /// `null` in contexts without map data, in which case terrain caps are skipped.
-  final Map<String, TileMapResult>? tileMapByRegion;
-
-  /// When set, avoids linear scans for Minor/Tribe checks in purchase-land
-  /// prevalidation (Refs #2394).
-  final DiplomacyFactionMembership? factionMembership;
-
-  /// True when the unit may perform civilian work in a Minor/Tribe province
-  /// with embassy + tech (validator-specific).
-  final bool Function(String unitType, String? provinceOwnerId)
-  civilianEmbassyWorkAllowed;
-}
-
-/// Returns a rejection, or `null` when this precheck passes.
-typedef WorkOrderTargetPrecheck =
-    OrderValidationResult? Function(
-      WorkOrderTargetPrecheckContext ctx,
-      WorkOrder o,
-      String? targetProvinceId,
-      String? provinceOwnerId,
-      String unitType,
-    );
-
-/// Targets that run dedicated territory rules in [workOrderTargetPrechecks] and
-/// must not also hit the default non-explorer foreign-province check.
-const Set<String> kWorkTargetsSkippingDefaultForeignProvinceCheck = {
-  ...kWorkTargetsWithoutMaterialCost,
-  kWorkTargetBuildImprovement,
-  kWorkTargetUpgradeTown,
-};
-
-/// Shared empty-resource + mineral-prospected gate for purchase_land /
-/// build_improvement (Refs #3949 item 7).
-({String? resourceId, OrderValidationResult? rejection})
-_resourceOrMineralRejection(
-  WorkOrderTargetPrecheckContext ctx,
-  String tileKey, {
-  required String emptyResourceMessage,
-}) {
-  final resourceId = ctx.game.worldState.resourceAtTile(tileKey);
-  if (resourceId == null || resourceId.isEmpty) {
-    return (
-      resourceId: null,
-      rejection: OrderValidationResult.rejected(emptyResourceMessage),
-    );
-  }
-  final mineralRejection = rejectIfMineralTileNotProspected(
-    game: ctx.game,
-    playerId: ctx.playerId,
-    tileKey: tileKey,
-    resourceId: resourceId,
-  );
-  if (mineralRejection != null) {
-    return (resourceId: null, rejection: mineralRejection);
-  }
-  return (resourceId: resourceId, rejection: null);
-}
-
-/// Shared controlled-or-embassy civilian access gate (Refs #3949 item 7).
-OrderValidationResult? _rejectIfUncontrolledWithoutEmbassyWork(
-  WorkOrderTargetPrecheckContext ctx,
-  String tileKey, {
-  required String unitType,
-  required String? provinceOwnerId,
-  required String message,
-}) {
-  final controlled = isTileControlledByPlayer(
-    ctx.game,
-    ctx.playerId,
-    tileKey,
-  );
-  final embassyWork = ctx.civilianEmbassyWorkAllowed(
-    unitType,
-    provinceOwnerId,
-  );
-  if (controlled || embassyWork) return null;
-  return OrderValidationResult.rejected(message);
-}
-
-OrderValidationResult? precheckUpgradeTown(
-  WorkOrderTargetPrecheckContext ctx,
-  WorkOrder order,
-  String? targetProvinceId,
-  String? provinceOwnerId,
-  String unitType,
-) {
-  if (ctx.player.techUnlocked?[kTechIdNationalBureaucracy] != true) {
-    return OrderValidationResult.rejected(
-      'National Bureaucracy tech required for upgrade_town',
-    );
-  }
-  if (targetProvinceId != null) {
-    final province = ctx.game.worldState.tryGetProvince(targetProvinceId);
-    if (province != null &&
-        province.townDevelopmentLevel >= kTownDevelopmentLevelMax) {
-      return OrderValidationResult.rejected(
-        'Town development level already at maximum (4)',
-      );
-    }
-    if (province != null) {
-      final townKey = province.townTileKey;
-      if (townKey == null ||
-          townKey.isEmpty ||
-          townKey != order.targetTileKey) {
-        return OrderValidationResult.rejected(
-          'upgrade_town target must be the province town tile',
-        );
-      }
-    }
-  }
-  if (provinceOwnerId != null && provinceOwnerId != ctx.playerId) {
-    if (!isMinorOrTribe(
-      ctx.game,
-      provinceOwnerId,
-      factionMembership: ctx.factionMembership,
-    )) {
-      return OrderValidationResult.rejected(
-        'upgrade_town target must be an owned or Minor/Tribe province town',
-      );
-    }
-    final embassyRejection = rejectAtWarOrWithoutEmbassy(
-      ctx.game,
-      ctx.playerId,
-      provinceOwnerId,
-      atWarMessage: 'Cannot upgrade town: at war with that faction',
-      embassyMessage:
-          'Cannot upgrade town: embassy required with that Minor/Tribe',
-    );
-    if (embassyRejection != null) {
-      return embassyRejection;
-    }
-  }
-  return null;
-}
-
-OrderValidationResult? precheckCounterSpy(
-  WorkOrderTargetPrecheckContext ctx,
-  WorkOrder order,
-  String? targetProvinceId,
-  String? provinceOwnerId,
-  String unitType,
-) {
-  if (provinceOwnerId != ctx.playerId) {
-    return OrderValidationResult.rejected(
-      'counter_spy target must be your own province',
-    );
-  }
-  return null;
-}
-
-OrderValidationResult? precheckPurchaseLand(
-  WorkOrderTargetPrecheckContext ctx,
-  WorkOrder o,
-  String? targetProvinceId,
-  String? ownerId,
-  String unitType,
-) {
-  if (ownerId == null || ownerId == ctx.playerId) {
-    return OrderValidationResult.rejected(
-      'purchase_land target must be a Minor or Tribe province',
-    );
-  }
-  if (!isMinorOrTribe(
-    ctx.game,
-    ownerId,
-    factionMembership: ctx.factionMembership,
-  )) {
-    return OrderValidationResult.rejected(
-      'purchase_land target must be a Minor or Tribe province',
-    );
-  }
-  final embassyRejection = rejectAtWarOrWithoutEmbassy(
-    ctx.game,
-    ctx.playerId,
-    ownerId,
-    atWarMessage: 'Cannot purchase land: at war with that faction',
-    embassyMessage:
-        'Cannot purchase land: embassy required with that Minor/Tribe',
-  );
-  if (embassyRejection != null) {
-    return embassyRejection;
-  }
-  final resourceGate = _resourceOrMineralRejection(
-    ctx,
-    o.targetTileKey,
-    emptyResourceMessage: 'Tile has no resource',
-  );
-  if (resourceGate.rejection != null) return resourceGate.rejection;
-  final resourceId = resourceGate.resourceId;
-  if (resourceId == null) {
-    return OrderValidationResult.rejected('Tile has no resource');
-  }
-  final cost = purchaseLandCost(resourceId);
-  if (ctx.treasury < cost) {
-    return OrderValidationResult.rejected(
-      'Insufficient treasury for purchase_land (need $cost)',
-    );
-  }
-  final existingBuyer = ctx.game.worldState.purchaserOfTile(o.targetTileKey);
-  if (existingBuyer != null) {
-    return OrderValidationResult.rejected(
-      existingBuyer == ctx.playerId
-          ? 'You already own this tile'
-          : 'Tile already purchased by another power',
-    );
-  }
-  return null;
-}
-
-OrderValidationResult? precheckBuildImprovement(
-  WorkOrderTargetPrecheckContext ctx,
-  WorkOrder o,
-  String? targetProvinceId,
-  String? ownerId,
-  String unitType,
-) {
-  final accessRejection = _rejectIfUncontrolledWithoutEmbassyWork(
-    ctx,
-    o.targetTileKey,
-    unitType: unitType,
-    provinceOwnerId: ownerId,
-    message: 'Cannot build improvement in foreign or uncontrolled province',
-  );
-  if (accessRejection != null) return accessRejection;
-  final resourceGate = _resourceOrMineralRejection(
-    ctx,
-    o.targetTileKey,
-    emptyResourceMessage:
-        'Tile has no resource; build_improvement requires a resource on the tile',
-  );
-  if (resourceGate.rejection != null) return resourceGate.rejection;
-  final resourceId = resourceGate.resourceId;
-  if (resourceId == null) {
-    return OrderValidationResult.rejected(
-      'Tile has no resource; build_improvement requires a resource on the tile',
-    );
-  }
-  final currentLevel = ctx.game.worldState.improvementLevelAtTile(
-    o.targetTileKey,
-  );
-  if (currentLevel >= 4) {
-    return OrderValidationResult.rejected(
-      'Improvement level already at maximum (4)',
-    );
-  }
-  final techCap = extractionCapForResourceForUnlocked(
-    ctx.player.techUnlocked,
-    resourceId,
-  );
-  final terrain = terrainTypeForTileKey(
-    ctx.tileMapByRegion,
-    o.targetTileKey,
-  );
-  final effectiveCap = terrain == null
-      ? techCap
-      : clampExtractionCapForTerrain(techCap, resourceId, terrain);
-  if (terrain != null && effectiveCap < techCap && currentLevel + 1 > effectiveCap) {
-    return OrderValidationResult.rejected(
-      'Terrain caps $resourceId extraction at level $effectiveCap on this '
-      'terrain (scrub forest timber is hard-capped at level 1)',
-    );
-  }
-  if (currentLevel + 1 > effectiveCap) {
-    return OrderValidationResult.rejected(
-      'Insufficient tech for next improvement level on $resourceId '
-      '(extraction cap $effectiveCap; unlock gathering tech to raise the cap)',
-    );
-  }
-  return null;
-}
-
-OrderValidationResult? precheckExplorerConsulateInMinorTribe(
-  WorkOrderTargetPrecheckContext ctx,
-  WorkOrder order,
-  String? targetProvinceId,
-  String? provinceOwnerId,
-  String unitType,
-) {
-  return rejectExplorerWithoutConsulateInMinorTribeProvince(
-    game: ctx.game,
-    playerId: ctx.playerId,
-    unitType: unitType,
-    workTarget: order.target,
-    provinceOwnerId: provinceOwnerId,
-    factionMembership: ctx.factionMembership,
-  );
-}
-
-OrderValidationResult? precheckDefaultForeignProvince(
-  WorkOrderTargetPrecheckContext ctx,
-  WorkOrder order,
-  String? targetProvinceId,
-  String? provinceOwnerId,
-  String unitType,
-) {
-  if (isExplorerUnit(unitType) ||
-      kWorkTargetsSkippingDefaultForeignProvinceCheck.contains(order.target)) {
-    return null;
-  }
-  return _rejectIfUncontrolledWithoutEmbassyWork(
-    ctx,
-    order.targetTileKey,
-    unitType: unitType,
-    provinceOwnerId: provinceOwnerId,
-    message: 'Cannot work in foreign province',
-  );
-}
-
-OrderValidationResult? precheckDevExclusiveTileConflict(
-  WorkOrderTargetPrecheckContext ctx,
-  WorkOrder order,
-  String? targetProvinceId,
-  String? provinceOwnerId,
-  String unitType,
-) {
-  if (!isDevExclusiveUnitType(unitType) ||
-      !isDevExclusiveWorkTarget(order.target)) {
-    return null;
-  }
-  if (!ctx.devExclusiveTiles.contains(order.targetTileKey)) {
-    return null;
-  }
-  return OrderValidationResult.rejected(
-    'Tile already has development or purchase work for this player',
-  );
-}
+export 'work_order_target_prechecks_explore.dart';
+export 'work_order_target_prechecks_improvement.dart';
+export 'work_order_target_prechecks_purchase.dart';
+export 'work_order_target_prechecks_shared.dart';
 
 /// Map dispatch for target-specific validation before generic work rules.
 final Map<String, WorkOrderTargetPrecheck> workOrderTargetPrechecks = {
